requirements.txt
# Put the additional Python libs you need here
requests==2.25.1
azure-servicebus==7.1.0
